requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
django-filter==23.4
djangorestframework==3.14.0

